fixed up folder layout but overall looks good
