+++ src/app/admin/courses/[courseId]/page.tsx
@@ -1,8 +1,13 @@
-export default async function Page({
-  params,
-}: {
-  params: Promise<{ courseId: string }>;
+
+export default async function CourseIdPage({ 
+  params 
+}: { 
+  params: Promise<{ courseId: string}> 
 }) {
-  const { courseId } = await params;
-  return <div>This is the admin course page for {courseId}</div>;
+  const { courseId } = await params; 
+  const id = Number(courseId);
+  
+  return <div className="">
+    Course ID: {id}
+  </div>
 }
