+++ src/app/admin/layout.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import AdminNavbar from "@/components/AdminNavbar";
+import AdminNavbar from "@/components/admin/AdminNavbar";
 import { Toaster } from "sonner"
