+++ src/lib/validations/course.ts
@@ -1,5 +1,4 @@
 import { z } from "zod";
-
 export const courseSchema = z.object({
     title: z.string().min(3, "Title must be at least 3 characters long").trim(),
     description: z.string().max(500, "Description is too long").trim().optional().or(z.literal("")),
@@ -10,8 +9,3 @@ export const courseSchema = z.object({
 
 });
 export type CourseFormValues = z.infer<typeof courseSchema>;
-
-export const tagSchema = z.object({
-    tagName: z.string().min(3, "Tag must be at least 3 characters long.").max(10, "Tag must be less than or equal to 10 characters long."), 
-});
-export type TagFormValues = z.infer<typeof tagSchema>;
